+++ src/oneid/oneid-lambda-cache-updater/src/main/java/it/pagopa/oneid/service/CacheUpdaterServiceImpl.java
@@ -5,6 +5,8 @@ import io.quarkus.logging.Log;
 import it.pagopa.oneid.connector.CloudWatchConnector;
 import it.pagopa.oneid.common.connector.CacheConnector;
 import it.pagopa.oneid.common.model.Client;
+import it.pagopa.oneid.common.utils.dynamodb.DynamoStreamService;
+import it.pagopa.oneid.common.utils.dynamodb.RecordUtils;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import java.util.List;