+++ src/oneid/oneid-lambda-client-publisher/src/main/java/it/pagopa/oneid/service/ClientServiceImpl.java
@@ -1,0 +1,38 @@
+package it.pagopa.oneid.service;
+
+import io.quarkus.runtime.Startup;
+import it.pagopa.oneid.common.connector.ClientConnector;
+import it.pagopa.oneid.common.model.Client;
+import it.pagopa.oneid.common.model.ClientFE;
+import it.pagopa.oneid.common.utils.logging.CustomLogging;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
+import jakarta.validation.constraints.NotBlank;
+import java.util.ArrayList;
+import java.util.Optional;
+import java.util.stream.Collectors;
+
+@ApplicationScoped
+@CustomLogging
+@Startup
+public class ClientServiceImpl implements ClientService {
+
+  private final ClientConnector clientConnector;
+
+  @Inject
+  ClientServiceImpl(ClientConnector clientConnector) {
+    this.clientConnector = clientConnector;
+  }
+
+  @Override
+  public Optional<ArrayList<ClientFE>> getAllClientsInformation() {
+    Optional<ArrayList<Client>> clients = clientConnector.findAll();
+    return clients.map(cl ->
+        cl.stream()
+          .filter(Client::isActive)
+            .map(ClientFE::new)
+            .collect(Collectors.toCollection(ArrayList::new)
+            )
+    );
+  }
+}