+++ src/infra/CHANGELOG.md
@@ -1,5 +1,17 @@
 # infra
 
+## 2.11.0
+
+### Minor Changes
+
+- 2f9f8ee: add lambda client publisher implementation and infra, update common with dynamo event deserialization logic and depshas
+
+## 2.10.0
+
+### Minor Changes
+
+- 363f4ad: enable cache solution in all environments, remove local cache in ecs and use distributed cache, add mterics and alarms for ecs and lambda cache updeter
+
 ## 2.9.0
 
 ### Minor Changes