+++ src/oneid/oneid-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # oneid-common
 
+## 1.6.1
+
+### Patch Changes
+
+- 2f9f8ee: add lambda client publisher implementation and infra, update common with dynamo event deserialization logic and depshas
+
 ## 1.6.0
 
 ### Minor Changes