+++ src/oneid/oneid-lambda-cache-updater/CHANGELOG.md
@@ -1,5 +1,17 @@
 # oneid-lambda-cache-updater
 
+## 0.3.1
+
+### Patch Changes
+
+- 2f9f8ee: add lambda client publisher implementation and infra, update common with dynamo event deserialization logic and depshas
+
+## 0.3.0
+
+### Minor Changes
+
+- 363f4ad: enable cache solution in all environments, remove local cache in ecs and use distributed cache, add mterics and alarms for ecs and lambda cache updeter
+
 ## 0.2.0
 
 ### Minor Changes